+++ tardis/tardis_portal/tasks.py
@@ -12,15 +12,9 @@ from django.db import transaction
 from django.db.models import Q
 from django.template import Context
 
-from tardis.tardis_portal.models import DataFile
-from tardis.tardis_portal.models import DataFileObject
-from tardis.tardis_portal.models import Dataset
-from tardis.tardis_portal.models import StorageBox
 from tardis.tardis_portal.staging import get_staging_url_and_size
 from tardis.tardis_portal.email import email_user
 
-# import tasks from elsewhere here so they are registered
-import tardis.tardis_portal.storage.squashfs
 
 # Ensure filters are loaded
 try:
@@ -37,24 +31,12 @@ except Exception:
 
 @task(name="tardis_portal.verify_dfos", ignore_result=True)
 def verify_dfos(dfos=None, **kwargs):
+    from tardis.tardis_portal.models import DataFileObject
     dfos_to_verify = dfos or DataFileObject.objects\
                                            .filter(verified=False)
+    kwargs['transaction_lock'] = kwargs.get('transaction_lock', True)
     for dfo in dfos_to_verify:
-        verify_dfo.apply_async(args=[dfo.id], **kwargs)
-
-
-@task(name="tardis_portal.verify_dfo", ignore_result=True)
-def verify_dfo(dfo_id, only_local=False, reverify=False):
-    '''
-    verify task
-    allowemtpychecksums is false for auto-verify, hence the parameter
-    '''
-    # Use a transaction for safety
-    with transaction.atomic():
-        # Get dfo locked for write (to prevent concurrent actions)
-        dfo = DataFileObject.objects.select_for_update().get(id=dfo_id)
-        if reverify or not dfo.verified:
-            dfo.verify()
+        dfo_verify.apply_async(args=[dfo.id], **kwargs)
 
 
 @task(name='tardis_portal.ingest_received_files', ignore_result=True)
@@ -63,6 +45,7 @@ def ingest_received_files():
     finds all files stored in temporary storage boxes and attempts to move
     them to their permanent home
     '''
+    from tardis.tardis_portal.models import StorageBox
     ingest_boxes = StorageBox.objects.filter(Q(attributes__key='type'),
                                              Q(attributes__value='receiving'),
                                              ~Q(master_box=None))
@@ -135,6 +118,7 @@ def create_staging_datafiles(files, user_id, dataset_id, is_secure):
 
 @task(name="tardis_portal.create_staging_datafile", ignore_result=True)
 def create_staging_datafile(filepath, username, dataset_id):
+    from tardis.tardis_portal.models import DataFile, Dataset
     dataset = Dataset.objects.get(id=dataset_id)
 
     url, size = get_staging_url_and_size(username, filepath)
@@ -163,3 +147,91 @@ def cache_done_notify(results, user_id, site_id, ct_id, obj_ids):
     }
     email_user(subject, 'cache_done_email', context, user)
     return "all done"
+
+
+# "method tasks"
+# StorageBox
+@task(name="tardis_portal.storage_box.copy_files", ignore_result=True)
+def sbox_copy_files(sbox_id, dest_box_id=None, *args, **kwargs):
+    from tardis.tardis_portal.models import StorageBox
+    sbox = StorageBox.objects.get(id=sbox_id)
+    if dest_box_id is not None:
+        dest_box = StorageBox.objects.get(id=dest_box_id)
+    else:
+        dest_box = None
+    return sbox.copy_files(dest_box=dest_box, *args, **kwargs)
+
+
+@task(name="tardis_portal.storage_box.move_files", ignore_result=True)
+def sbox_move_files(sbox_id, dest_box_id=None, *args, **kwargs):
+    from tardis.tardis_portal.models import StorageBox
+    sbox = StorageBox.objects.get(id=sbox_id)
+    if dest_box_id is not None:
+        dest_box = StorageBox.objects.get(id=dest_box_id)
+    else:
+        dest_box = None
+    return sbox.move_files(dest_box=dest_box, *args, **kwargs)
+
+
+@task(name='tardis_portal.storage_box.copy_to_master')
+def sbox_copy_to_master(sbox_id, *args, **kwargs):
+    from tardis.tardis_portal.models import StorageBox
+    sbox = StorageBox.objects.get(id=sbox_id)
+    return sbox.copy_to_master(*args, **kwargs)
+
+
+@task(name='tardis_portal.storage_box.move_to_master')
+def sbox_move_to_master(sbox_id, *args, **kwargs):
+    from tardis.tardis_portal.models import StorageBox
+    sbox = StorageBox.objects.get(id=sbox_id)
+    return sbox.move_to_master(*args, **kwargs)
+
+
+# DataFile
+@task(name="tardis_portal.cache_datafile", ignore_result=True)
+def df_cache_file(df_id):
+    from tardis.tardis_portal.models import DataFile
+    df = DataFile.objects.get(id=df_id)
+    return df.cache_file()
+
+
+# DataFileObject
+@task(name='tardis_portal.dfo.move_file', ignore_result=True)
+def dfo_move_file(dfo_id, dest_box_id=None):
+    from tardis.tardis_portal.models import DataFileObject, StorageBox
+    dfo = DataFileObject.objects.get(id=dfo_id)
+    if dest_box_id is not None:
+        dest_box = StorageBox.objects.get(id=dest_box_id)
+    else:
+        dest_box = None
+    return dfo.move_file(dest_box)
+
+
+@task(name='tardis_portal.dfo.copy_file', ignore_result=True)
+def dfo_copy_file(dfo_id, dest_box_id=None, *args, **kwargs):
+    from tardis.tardis_portal.models import DataFileObject, StorageBox
+    dfo = DataFileObject.objects.get(id=dfo_id)
+    if dest_box_id is not None:
+        dest_box = StorageBox.objects.get(id=dest_box_id)
+    else:
+        dest_box = None
+    return dfo.copy_file(dest_box=dest_box, *args, **kwargs)
+
+
+@task(name='tardis_portal.dfo.cache_file', ignore_result=True)
+def dfo_cache_file(dfo_id):
+    from tardis.tardis_portal.models import DataFileObject
+    dfo = DataFileObject.objects.get(id=dfo_id)
+    return dfo.cache_file()
+
+
+@task(name="tardis_portal.dfo.verify", ignore_result=True)
+def dfo_verify(dfo_id, *args, **kwargs):
+    from tardis.tardis_portal.models import DataFileObject
+    # Get dfo locked for write (to prevent concurrent actions)
+    if kwargs.pop('transaction_lock', False):
+        with transaction.atomic():
+            dfo = DataFileObject.objects.select_for_update().get(id=dfo_id)
+            return dfo.verify(*args, **kwargs)
+    dfo = DataFileObject.objects.get(id=dfo_id)
+    dfo.verify(*args, **kwargs)