+++ docs/changes.rst
@@ -13,6 +13,8 @@ in development
   than 256 characters.
 * Changed constraints on the instrument model; facility and instrument name are
   now unique together
+* changed method tasks to task functions, pre-empting the removal of methods
+  tasks in new celery versions
 
 
 3.6 - 16 March 2015